move github response type definitions to a dedicated file
